Automatic macro hygiene based on "scope layers"
When expanding macros

* Any identifiers passed to the macro are tagged with the scope layer they were
  defined within.
* A new unique scope layer is generated for the macro invocation, and any names
  in the syntax produced by the macro are tagged with this layer.

Subsequently, the `(name,scope_layer)` pairs are used rather than plain
names when resolving bindings.

This simple scheme is powerful enough to allow macros to call macros and
this composition to work by default. It even allows macros to emit
recursive calls to themselves in the code they return. More
experimentation is needed to determine whether it's a complete fit for
Julia's needs.