Reimplement quoted syntax interpolation
Quoted syntax containing `$` interpolations can be expanded into a tree
of nested expression generating code, but this generates a lot of code
for nested expressions; effectively it's a "loop unrolling approach".

Alternatively, we can just inert-quote the `$`-containing expression as
a single expression in the source, separately strip out the interpolated
values into their own expressions, and emit a call to a runtime function
to do the interpolation.

We take the second approach here; it results in much cleaner expanded
expressions and keeps the original quoted syntax in a single data
structure.