Desugaring of `struct` type definitions
Still todo:
* inner constructors
* outer constructors
* doc binding

Also included here is `K"alias_binding"` - a more general replacement
for the `outerref` used in flisp lowering. `alias_binding` allows one to
allocate a binding early during desugaring and make this binding an
alias for a given name. Bindings don't participate in scope resolution,
so this allows us to bypass the usual scoping rules. For example, to
refer to a global struct_name from an outer scope, but within an inner
scope where the identifier struct_name is bound to a local variable.
(We could also replace outerref by generating a new scope_layer and
perhaps that would be simpler?)