Use scope layers for struct lowering rather than `K"alias_binding"`
As much as alias_binding is a neat idea, it seems like using a scope
layer to distinguish the global vs local bindings might be good enough
and allow us to remove the alias_binding concept. As a side effect, this
may allow us to avoid needing support arbitrary bindings in some early
lowering code.