`@ nospecialize` for function arguments
I've chosen to attach the nospecialize metadata here as an attribute on
the function argument names. This ensures it travels with the function
arguments, without otherwise disturbing the AST.

TODO: `@ nospecialise` within the function body is not done yet - it's
not very natural in this scheme but I guess we should still be able to
recognize it during scope analysis and turn it into a tombstone, moving
the metadata onto the lambda's argument bindings.