Get basic closure conversion somewhat working
Implement basic creation of closure types with all variables boxed.
Variables outside the closure are not handled correctly yet.

Desugaring of functions which become closures is tricky because we don't
yet know that they're closures and we might move their method
definitions to top level. To aid with this, introduce three intermediate
forms:
* `K"function_decl"` to create a local or global binding for the
  function name,
* `K"function_type"` to evaluate the type (either global generic
  function or closure converted struct)
* `K"method_defs"` to encapsulate code which creates methods and might
  be moved to top level. (Just using `K"method"` isn't ideal for this
  because we might want to evaluate some intermediate forms and reuse
  them for multiple method definitions for convenience and to avoid
  multiple evaluation of expressions with side effects. Eg when lowering
  functions with default positional args to multiple method defs.)