Store binding provenance in BindingInfo
Here we give every binding a source reference (stored as node_id) so
that if we only have the BindingInfo we can still insert it into the AST
with sensible proveneance. For example, for `@locals`, we're required to
reconstruct a list of bindings without any source construct to hang them
off other than the `@locals` invocation. With this change we can include
the original location where the binding was defined. This also fixes
some pretty misleading printing of the `#self#` binding for closures.

Additional cleanup:
* Regroup all binding-related stuff into its own source file
* Remove alias bindings. These seemed like a good idea at the time but
  so far, there's other ways to achieve the same thing.
* Rename new_mutable_var to new_local_binding