Ensure `TypeVar` definitions accompany method defs moved to top level
The `K"method_defs"` needs to group a preamble+sets of methods for a
function together so that typevars can be defined in the preamble and
the whole thing moved to top level as one piece, if necessary, during
closure conversion. Previous refactoring had broken that - reinstated it
here.

There's some repetition of the typevars here for certain cases like
kwcall overloads vs the actual function overloads, vs the separate kw
body function - we can't tell during desugaring whether a function will
become a closure or not so it's possible we might need to run these
TypeVar initializations twice.