add precompilation statements to improve first-time-to-lower latency (#14)
Exercises the complete lowering pipeline during precompilation with a
sample code thunks to create method and type definitions.

This precompilation is enabled by default but can be disabled by setting
the `JULIA_LOWERING_PRECOMPILE` environment variable to false-cy, which
may be useful during development.

Ideally we'd like to use Preferences.jl for this configuration, but
my understanding is that JuliaLowering aims to be self-contained and
should not add dependencies other than JuliaSyntax.