Fixes for opaque closure method lowering
opaque_closure_method has special non-evaluated semantics for its
argument list, so we need some special non-quoted conversion here for
the functionloc argument and the name argument when it's set to
the global ref Core.nothing.

To fix the globalref, I've chosen to translate `"core"::K"nothing"` into
a literal `nothing` in general - this is consistent with how ast.c
translates the special flisp `(null)` form.