Separate `JuliaLowering.eval()` from `Core.eval()`
Overloading `Core.eval()` with `SyntaxTree` was a cute trick but causes
invalidations in the include machinery (Base.IncludeInto) and doesn't
really integrate JuliaLowering properly because it doesn't let us
implement `Core.eval(::Module, ::Expr)`. Instead we need hooks for this.

For now, implementing `eval` as a separate function helps quite a lot
with precompile time - about a 15% reduction on my machine.