Take 2 on use of `global x` in value position
Disallows "use" of the value of `global x`, except in tail position in
top level thunks - in those cases, return `nothing` so that `global x`
can be used in the value position of a top level thunk. This is normally
harmless as one cannot observe this value, except in special
circumstances - namely the return value of `eval()` (and things which
call eval, such as `include()`).

While we're thinking about this, also disallow a bare `local x` in a top
level thunk because this cannot have useful side effects and is just
confusing when it occurs outside a block construct. (This is not
currently disallowed for `local` arising from macro expansions because
it's not an obvious user error in that case. It could possibly arise as
a valid macro expansion of a trivial case, for some macros?)