Fixes for global and local decls in value position (#67)
Disallows "use" of the value of `global x`, except in tail position in
top level thunks - in those cases, return `nothing` so that `global x`
can be used nested within of a top level thunk as long as it's "statement like".
This is normally harmless as one cannot observe the value from tail position
except in special circumstances - namely the return value of `eval()`
(and things which call eval, such as `include()`). It's a conservative relaxation
of the current rules where globals in value position are only allowed in top
level code if they're not nested in other syntax.

While we're thinking about this, also disallow a bare `local x` in a top
level thunk because this cannot have useful side effects and is just
confusing when it occurs outside a block construct. (This is not
currently disallowed for `local` arising from macro expansions because
it's not an obvious user error in that case. It could possibly arise as
a valid macro expansion of a trivial case, for some macros?)