`expr_to_syntaxtree`: Quoted symbol fixes (#78)
Fix #54.  QuoteNode was previously converted to `(quote <contents>)` which lowers to a runtime call.  
- For plain symbol contents, this was too dynamic for ccall to handle, and optimizable in any case.  
- For Expr contents, I don't think using non-interpolating quote was correct in the first place.
- For LineNumberNode contents, we wrap with a QuoteNode on the way out, so we need to unwrap it on the way in.

I still need to think more about what the various forms of quoting mean when Expr and SyntaxTree are mixed, but this is at least an upgrade in number of passing tests.  I've added some cases from things I broke along the way.

Other fixes:
- Bool isa Integer, so prevent `expr_to_syntaxtree` from giving booleans `K"Integer"` kind
- enable logging of failures from `expr_to_syntaxtree` in the core lowering hook

---------

Co-authored-by: Claire Foster <aka.c42f@gmail.com>