Compiler frontend API and sketch for Core integration
Here I've generalized and refactored the incremental lowering code from
PR #84, splitting out the interface from the implementation and
factoring the parts which should go into `Core` into a `_Core` module.

`_Core.eval()` is implemented in terms of a new `_Core.simple_eval()`
which can evaluate only `CodeInfo` and the new types
`TopLevelCodeIterator`, `BeginModule` and `EndModule`. Evaluation of
`TopLevelCodeIterator` is implemented in terms of the incremental
lowering interface which must be provided by a compiler frontend.

A compiler frontend (subtype of `CompilerFrontend`) must come with
implementations of the functions `lower_init` and `parseall` and
`lower_step` must be being defined on the return type of `lower_init`.
Having a type for the frontend solves the issue of how to tie parsing
and lowering together without needing to convert to `Expr`.

`parseall(frontend, code, ...)` returns a syntax tree of the preferred
expression type for the frontend which can be fed into `lower_init`.
This function is just a placeholder - it needs to be generalized to
allow `Meta.parse()`, `Meta.parseatom()` and `Meta.parseall()` to be
implemented in terms of it.

Parser and lowering diagnostics are handled by throwing an exception
from the frontend but we may want a fuller diagnostics API instead.