Use relative import paths in tests
For vendoring into Base we need to avoid absolute import paths as in
`using JuliaLowering` and `using JuliaSyntax` in the test files as
neither of these packages will be top level modules. Thus, replace all
occurrences of these with relative import paths except for one central
location (currently in util.jl) which can be easily adjusted.